Moved createListItem to list-componenet.js
