+++ test/list-component.test.js
@@ -1,18 +1,8 @@
+import { createListItem } from '../src/list-component.js';
+
 const test = QUnit.test;
 
 QUnit.module('List component');
-
-function createListItem(queen) {
-    const html = /*html*/`
-        <li>
-            <img src="${queen.image_url}" alt="Image of ${queen.name}">
-            <a href="queen-detail.html?id=${queen.id}">${queen.name}</a>            
-        </li>
-    `;
-    const template = document.createElement('template');
-    template.innerHTML = html;
-    return template.content;
-}
 
 test('create list item', function(assert) {
     // arrange
